requirements.txt
# Scalping Bot - Powell ICT Detector (NQ/ES futures)
# Data: yfinance for NQ=F, ES=F
pandas>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
yfinance>=0.2.28
ccxt>=4.0.0
python-dateutil>=2.8.2
pyyaml>=6.0
python-telegram-bot>=20.0
loguru>=0.7.0
